增加作者

+++ bs4/mm_jpg.py
@@ -24,28 +24,40 @@ def download_list_pics(n,piclist,root):
 
 #下载图片，并写入文件
 def download_pics(n,i,url,root,name):
+    offset = url
     # 请求头
     headers = {'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
                'Accept-Encoding': 'gzip, deflate, sdch',
                'Accept-Language': 'zh-CN,zh;q=0.8',
                'Connection': 'keep-alive',
                'Host': 'www.mmjpg.com',
-               'Referer':'http://img.mmjpg.com/2016/767/30.jpg',
+               'Referer': offset,
                'Upgrade-Insecure-Requests': '1',
                'User-Agent': 'Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/55.0.2883.87 Safari/537.36'}
 
-    ir = session.get(url, headers=headers)
+
     path = root + '\\'+str(name)+'.jpg'
-    # print('下载url:', url)
-    # print('下载status_code:', ir.status_code)
-    # # print('下载Location:',ir.headers["Location"])
-    # print('下载history:', ir.history)
-    if ir.status_code == 200:
-        print('下载ok:',n,' ',i,'' ,path)
-        with open(path, 'wb') as f:
-            f.write(ir.content)
+    # print('下载:', url)
+    isExists = os.path.exists(path)
+    # ir = session.get(url, headers=headers)
+    # if ir.status_code == 200:
+    #     print('下载ok:', n, ' ', i, '', path)
+    #     with open(path, 'wb') as f:
+    #         f.write(ir.content)
+    # else:
+    #     print('下载ng:', n, ' ', i, '', path)
+
+    if  isExists:
+        ir = session.get(url, headers=headers)
+        if ir.status_code == 200:
+            print('下载ok:', n, ' ', i, '', path)
+            with open(path, 'wb') as f:
+                f.write(ir.content)
+        else:
+            print('下载ng:', n, ' ', i, '', path)
     else:
-        print('下载ng:', n, ' ', i, '', path)
+        print('文件存在不下载:', n, ' ', i, '', path)
+
 
 
 def get_html(url):
@@ -58,30 +70,15 @@ def get_html(url):
     except:
         return "get_html someting wrong"
 
+def get_ji_page_content(url):
 
-def get_content(url):
     html = get_html(url)
     soup = BeautifulSoup(html, 'lxml')
-    yi_list = soup.find('div', class_= 'pic')
-
-
-
-    # 找到列表
-    title_list = yi_list.find_all('li')
-    total = yi_list.find_all('em', class_= 'info')
-    # print('title_list:', title_list)
-    # total = yi_list.find_all('a', class_='last')
-    # print('total:', total)
-
-
-
-    count=0
+    yilist = soup.find('div', class_= 'pic')
+    title_list = yilist.find_all('li')
+    count = 0
     list=[]
     for mylist in  title_list:
-        # print('mylist:', mylist)
-        # print('mylist2:', mylist.find('a'))
-        # print('mylist3:', mylist.find('span'))
-        # <li><a href="http://www.mmjpg.com/mm/767" target="_blank"><img alt="绝色诱惑!身材超正的90后极品乳神于姬" height="330" src="http://img.mmjpg.com/small/2016/767.jpg" width="220"/></a><span class="title"><a href="http://www.mmjpg.com/mm/767" target="_blank">绝色诱惑!身材超正的90后极品乳神于姬</a></span><span>10-09 发布</span><span class="view">浏览(2917226)</span></li>
         name = mylist.find('a').img['alt']
         picurl = mylist.find('a')['href']
         count = count + 1
@@ -94,7 +91,30 @@ def get_content(url):
             'href': picurl,
         }
         list.append(vid)
-    return list;
+    # print('图集:', url,' ',count)
+    return list
+
+def get_content(url):
+    html = get_html(url)
+    soup = BeautifulSoup(html, 'lxml')
+    pagelist = soup.find('div', class_='page')
+
+    page_list = pagelist.find_all('em')
+    tt = page_list[0].text
+    count = tt[1:len(tt)-1]
+    print('下载图集页数:', tt,' ',count,' ',url)
+
+    downpagelist = []
+    for i in range(1,int(count)+1):
+      mypicurl = url + '/' + str(i)
+      # print('下载图集地址:', mypicurl)
+      downpagelist.append(mypicurl)
+    tag_list = []
+    for i  in range(0,len(downpagelist)):
+        list = get_ji_page_content(downpagelist[i]);
+        tag_list = tag_list + list
+    print('下载图集总数:', len(tag_list))
+    return tag_list
 
 def get_page_content(url):
     pichtml = get_html(url)
@@ -134,7 +154,6 @@ def get_page_content(url):
         alt = soup.find('div', class_='content').find('a').find('img')['alt']
         # print('title:', title)
         # print('src:', downpath, ' alt:', alt)
-
         vid = {
             'title': title,
             'src': downpath,
@@ -171,8 +190,8 @@ def main():
     for i in range(0,len(taglist)):
         # print('title:', taglist[i]['title'])
         root_dir_2 = create_dir(root_dir +  taglist[i]['title']+ '\\') # 绝对路径
-        list = get_content(url)
-        print('图集个数:',len(list))
+        list = get_content(taglist[i]['href'])
+        # print('图集个数:',len(list))
         for i in range(0,len(list)):
             print('下载图集:',i+1,' ', list[i]['title'],list[i]['href'])
             piclist = get_page_content(list[i]['href'])
